+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/document/FileCollector.java
@@ -10,7 +10,6 @@ import java.net.URI;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystem;
-import java.nio.file.FileSystemAlreadyExistsException;
 import java.nio.file.FileSystemNotFoundException;
 import java.nio.file.FileSystems;
 import java.nio.file.FileVisitOption;
@@ -315,30 +314,6 @@ public final class FileCollector implements AutoCloseable {
         } else {
             reporter.error("Not a file or directory {0}", file);
             return false;
-        }
-    }
-
-    /**
-     * Opens a zip file and returns a FileSystem for its contents, so
-     * it can be explored with the {@link Path} API. You can then call
-     * {@link #addFile(Path)} and such. The zip file is registered as
-     * a resource to close at the end of analysis.
-     *
-     * @deprecated Use {@link #addZipFileWithContent(Path)} instead.
-     */
-    @Deprecated
-    public FileSystem addZipFile(Path zipFile) {
-        if (!Files.isRegularFile(zipFile)) {
-            throw new IllegalArgumentException("Not a regular file: " + zipFile);
-        }
-        URI zipUri = URI.create("jar:" + zipFile.toUri());
-        try {
-            FileSystem fs = FileSystems.newFileSystem(zipUri, Collections.<String, Object>emptyMap());
-            resourcesToClose.add(fs);
-            return fs;
-        } catch (FileSystemAlreadyExistsException | ProviderNotFoundException | IOException e) {
-            reporter.errorEx("Cannot open zip file " + zipFile, e);
-            return null;
         }
     }